Big Update: Reorganization
I've organized the assets folder to put everything in its own folder.
Also, took out all of the messy junk in the assets folder and now our
two 'sample' scenes are actual panoramas Eric took.  In addition, you
can also open the map menu with the top left corner button, it'll freeze
the screen, and you can click the 2 to navigate to the other scene.